code cleanup + test player
